requirements.txt
slackclient>=2.3.1
